+++ app/managers/ContainerManager.php
@@ -2,28 +2,28 @@
 
 namespace App\Managers;
 
-use App\Constants\Container\CustomMetadataTypes;
-use App\Constants\Container\Processes\InvoiceCustomMetadata;
-use App\Constants\Container\StandaloneProcesses;
 use App\Constants\Container\SystemGroups;
 use App\Constants\ContainerStatus;
 use App\Core\Caching\CacheNames;
 use App\Core\DatabaseConnection;
 use App\Core\DB\DatabaseManager;
-use App\Core\FileManager;
+use App\Core\DB\DatabaseMigrationManager;
 use App\Core\HashManager;
-use App\Entities\ContainerDatabaseEntity;
 use App\Entities\ContainerEntity;
 use App\Exceptions\AException;
 use App\Exceptions\GeneralException;
 use App\Exceptions\NonExistingEntityException;
-use App\Helpers\ContainerCreationHelper;
 use App\Logger\Logger;
 use App\Repositories\Container\GroupRepository;
 use App\Repositories\ContainerRepository;
 use App\Repositories\ContentRepository;
 use App\Repositories\UserRepository;
 
+/**
+ * ContainerManager is used for managing containers
+ * 
+ * @author Lukas Velek
+ */
 class ContainerManager extends AManager {
     public ContainerRepository $containerRepository;
     private DatabaseManager $dbManager;
@@ -50,6 +50,16 @@ class ContainerManager extends AManager {
         return 'sd_db_' . $containerId . '_' . HashManager::createHash(8, false);
     }
 
+    /**
+     * Creates a new container
+     * 
+     * @param string $title Container title
+     * @param string $description Container description
+     * @param string $callingUserId Calling user ID
+     * @param int $environment Container environment
+     * @param bool $canShowReferent Can show referent
+     * @param int $status Container status
+     */
     public function createNewContainer(string $title, string $description, string $callingUserId, int $environment, bool $canShowReferent, int $status = ContainerStatus::NEW) {
         $containerId = $this->createId(EntityManager::CONTAINERS);
         $databaseName = $this->generateContainerDatabaseName($containerId);
@@ -88,386 +98,59 @@ class ContainerManager extends AManager {
         return $containerId;
     }
 
+    /**
+     * Creates a new container
+     * 
+     * This method is used asynchronously by a background service.
+     * 
+     * @param string $containerId Container ID
+     */
     public function createNewContainerAsync(string $containerId) {
         $container = $this->getContainerById($containerId, true);
 
         try {
-            //$this->dbManager->createNewDatabase($container->getDefaultDatabase()->getName());
-
-            $this->createNewContainerTables($container->getDefaultDatabase()->getName());
-            $this->createContainerTablesIndexes($container->getDefaultDatabase()->getName());
-            $this->updateContainerDbSchema($container->getDefaultDatabase()->getName(), $containerId);
-            
-            $exceptions = [];
-            $this->insertNewContainerDefaultDataAsync($containerId, $container, $container->getDefaultDatabase()->getName(), $exceptions);
+            $this->runContainerDatabaseMigrations($container->getDefaultDatabase()->getName(), $containerId);
         } catch(AException $e) {
             throw $e;
         }
     }
 
-    private function insertNewContainerDefaultDataAsync(string $containerId, ContainerEntity $container, string $dbName, array &$exceptions) {
+    /**
+     * Runs container database migrations
+     * 
+     * @param string $dbName Container database name
+     * @param string $containerId Container ID
+     */
+    public function runContainerDatabaseMigrations(string $dbName, string $containerId) {
         try {
             $conn = $this->dbManager->getConnectionToDatabase($dbName);
         } catch(AException $e) {
-            throw new GeneralException('Could not establish connection to the container database.');
+            throw new GeneralException('Could not establish connection to the container database. Reason: ' . $e->getMessage(), $e);
         }
 
-        $users = $this->groupManager->getGroupUsersForGroupTitle($container->getTitle() . ' - users');
+        $migrationManager = new DatabaseMigrationManager($this->masterConn, $conn, $this->logger);
+        $migrationManager->setContainer($containerId);
 
-        $groupIds = [];
-        foreach(SystemGroups::getAll() as $value => $text) {
-            $groupIds[$value] = $this->createIdCustomDb(EntityManager::C_GROUPS, $conn);
-        }
-
-        $folderIds = [
-            'Default' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_FOLDERS, $conn),
-            'Invoices' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_FOLDERS, $conn)
-        ];
-
-        $classIds = [
-            'Default' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_CLASSES, $conn),
-            'Invoices' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_CLASSES, $conn)
-        ];
-
-        $metadataIds = [
-            InvoiceCustomMetadata::COMPANY => $this->createIdCustomDb(EntityManager::C_CUSTOM_METADATA, $conn),
-            InvoiceCustomMetadata::SUM => $this->createIdCustomDb(EntityManager::C_CUSTOM_METADATA, $conn),
-            InvoiceCustomMetadata::INVOICE_NO => $this->createIdCustomDb(EntityManager::C_CUSTOM_METADATA, $conn),
-            InvoiceCustomMetadata::SUM_CURRENCY => $this->createIdCustomDb(EntityManager::C_CUSTOM_METADATA, $conn)
-        ];
-        
-        $data = [
-            [
-                'table' => 'document_classes',
-                'data' => [
-                    'classId' => $classIds['Default'],
-                    'title' => 'Default'
-                ]
-            ],
-            [
-                'table' => 'document_classes',
-                'data' => [
-                    'classId' => $classIds['Invoices'],
-                    'title' => 'Invoices'
-                ]
-            ],
-            [
-                'table' => 'document_class_group_rights',
-                'data' => [
-                    'rightId' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_CLASS_GROUP_RIGHTS, $conn),
-                    'groupId' => $groupIds[SystemGroups::ACCOUNTANTS],
-                    'classId' => $classIds['Invoices'],
-                    'canView' => 1,
-                    'canCreate' => 1,
-                    'canEdit' => 1
-                ]
-            ],
-            [
-                'table' => 'document_class_group_rights',
-                'data' => [
-                    'rightId' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_CLASS_GROUP_RIGHTS, $conn),
-                    'groupId' => $groupIds[SystemGroups::ALL_USERS],
-                    'classId' => $classIds['Default'],
-                    'canView' => 1,
-                    'canCreate' => 1
-                ]
-            ],
-            [
-                'table' => 'document_class_group_rights',
-                'data' => [
-                    'rightId' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_CLASS_GROUP_RIGHTS, $conn),
-                    'groupId' => $groupIds[SystemGroups::ADMINISTRATORS],
-                    'classId' => $classIds['Default'],
-                    'canView' => 1,
-                    'canCreate' => 1,
-                    'canEdit' => 1,
-                    'canDelete' => 1
-                ]
-            ],
-            [
-                'table' => 'document_folders',
-                'data' => [
-                    'folderId' => $folderIds['Default'],
-                    'title' => 'Default',
-                    'isSystem' => 1
-                ]
-            ],
-            [
-                'table' => 'document_folders',
-                'data' => [
-                    'folderId' => $folderIds['Invoices'],
-                    'title' => 'Invoices',
-                    'isSystem' => 1
-                ]
-            ],
-            [
-                'table' => 'document_folder_group_relation',
-                'data' => [
-                    'relationId' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_FOLDER_GROUP_RELATION, $conn),
-                    'folderId' => $folderIds['Default'],
-                    'groupId' => $groupIds[SystemGroups::ALL_USERS],
-                    'canView' => 1,
-                    'canCreate' => 1
-                ]
-            ],
-            [
-                'table' => 'document_folder_group_relation',
-                'data' => [
-                    'relationId' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_FOLDER_GROUP_RELATION, $conn),
-                    'folderId' => $folderIds['Default'],
-                    'groupId' => $groupIds[SystemGroups::ADMINISTRATORS],
-                    'canView' => 1,
-                    'canCreate' => 1,
-                    'canEdit' => 1,
-                    'canDelete' => 1
-                ]
-            ],
-            [
-                'table' => 'document_folder_group_relation',
-                'data' => [
-                    'relationId' => $this->createIdCustomDb(EntityManager::C_DOCUMENT_FOLDER_GROUP_RELATION, $conn),
-                    'folderId' => $folderIds['Invoices'],
-                    'groupId' => $groupIds[SystemGroups::ACCOUNTANTS],
-                    'canView' => 1,
-                    'canCreate' => 1,
-                    'canEdit' => 1
-                ]
-            ],
-            [
-                'table' => 'group_rights_standard_operations',
-                'data' => [
-                    'rightId' => $this->createIdCustomDb(EntityManager::C_GROUP_STANDARD_OPERATION_RIGHTS, $conn),
-                    'groupId' => $groupIds[SystemGroups::ADMINISTRATORS],
-                    'canShareDocuments' => 1,
-                    'canExportDocuments' => 1,
-                    'canViewDocumentHistory' => 1
-                ]
-            ],
-            [
-                'table' => 'archive_folders',
-                'data' => [
-                    'folderId' => $folderIds['Default'],
-                    'title' => 'Default',
-                    'isSystem' => 1
-                ]
-            ],
-            [
-                'table' => 'custom_metadata',
-                'data' => [
-                    'metadataId' => $metadataIds['Invoices_SumCurrency'],
-                    'title' => InvoiceCustomMetadata::SUM_CURRENCY,
-                    'guiTitle' => InvoiceCustomMetadata::toString(InvoiceCustomMetadata::SUM_CURRENCY),
-                    'type' => CustomMetadataTypes::SYSTEM_INVOICE_SUM_CURRENCY,
-                    'isRequired' => 1
-                ]
-            ],
-            [
-                'table' => 'custom_metadata',
-                'data' => [
-                    'metadataId' => $metadataIds['Invoices_Sum'],
-                    'title' => InvoiceCustomMetadata::SUM,
-                    'guiTitle' => InvoiceCustomMetadata::toString(InvoiceCustomMetadata::SUM),
-                    'type' => CustomMetadataTypes::NUMBER,
-                    'isRequired' => 1
-                ]
-            ],
-            [
-                'table' => 'custom_metadata',
-                'data' => [
-                    'metadataId' => $metadataIds['Invoices_InvoiceNo'],
-                    'title' => InvoiceCustomMetadata::INVOICE_NO,
-                    'guiTitle' => InvoiceCustomMetadata::toString(InvoiceCustomMetadata::INVOICE_NO),
-                    'type' => CustomMetadataTypes::TEXT,
-                    'isRequired' => 1
-                ]
-            ],
-            [
-                'table' => 'custom_metadata',
-                'data' => [
-                    'metadataId' => $metadataIds['Invoices_Company'],
-                    'title' => InvoiceCustomMetadata::COMPANY,
-                    'guiTitle' => InvoiceCustomMetadata::toString(InvoiceCustomMetadata::COMPANY),
-                    'type' => CustomMetadataTypes::SYSTEM_INVOICE_COMPANIES,
-                    'isRequired' => 1
-                ]
-            ],
-        ];
-
-        foreach($metadataIds as $title => $id) {
-            $data[] = [
-                'table' => 'document_folder_custom_metadata_relation',
-                'data' => [
-                    'relationId' => $this->createIdCustomDb(EntityManager::C_CUSTOM_METADATA_FOLDER_RELATION, $conn),
-                    'customMetadataId' => $id,
-                    'folderId' => $folderIds['Invoices']
-                ]
-            ];
-        }
-
-        $standaloneProcessIds = [];
-        foreach(StandaloneProcesses::getAll() as $key => $title) {
-            if(StandaloneProcesses::isDisabled($key)) continue;
-
-            $standaloneProcessIds[$key] = $this->createIdCustomDb(EntityManager::C_PROCESS_TYPES, $conn);
-
-            $data[] = [
-                'table' => 'process_types',
-                'data' => [
-                    'typeId' => $standaloneProcessIds[$key],
-                    'typeKey' => $key,
-                    'title' => $title,
-                    'description' => StandaloneProcesses::getDescription($key)
-                ]
-            ];
-        }
-
-        $data[] = [
-            'table' => 'process_metadata',
-            'data' => [
-                'metadataId' => $this->createIdCustomDb(EntityManager::C_PROCESS_CUSTOM_METADATA, $conn),
-                'typeId' => $standaloneProcessIds[StandaloneProcesses::INVOICE],
-                'title' => 'companies',
-                'guiTitle' => 'Companies',
-                'type' => CustomMetadataTypes::ENUM,
-                'isRequired' => '1'
-            ]
-        ];
-
-        foreach($groupIds as $value => $groupId) {
-            $data[] = [
-                'table' => 'groups',
-                'data' => [
-                    'groupId' => $groupId,
-                    'title' => $value
-                ]
-            ];
-        }
-
-        foreach($users as $userId) {
-            foreach($groupIds as $name => $groupId) {
-                $data[] = [
-                    'table' => 'group_users_relation',
-                    'data' => [
-                        'relationId' => $this->createIdCustomDb(EntityManager::C_GROUP_USERS_RELATION, $conn),
-                        'userId' => $userId,
-                        'groupId' => $groupId
-                    ]
-                ];
-            }
-        }
-
-        // DATA INSERT
-        foreach($data as $part) {
-            try {
-                $tableName = $part['table'];
-                $values = $part['data'];
-
-                $this->dbManager->insertDataToTable($tableName, $values, $dbName);
-            } catch(AException $e) {
-                $exceptions[$tableName] = $e;
-                continue;
-            }
-        }
+        $migrationManager->runMigrations(true);
     }
 
-    private function createNewContainerTables(string $dbName, array $tableDefinitions = []) {
-        if(empty($tableDefinitions)) {
-            $tables = ContainerCreationHelper::getContainerTableDefinitions();
-        } else {
-            $tables = $tableDefinitions;
-        }
-
-        foreach($tables as $name => $definition) {
-            if(!$this->dbManager->createTable($name, $definition, $dbName)) {
-                throw new GeneralException('Could not create database table.');
-            }
-        }
-    }
-
-    private function createContainerTablesIndexes(string $dbName, array $indexDefinitions = []) {
-        if(empty($indexDefinitions)) {
-            $indexes = ContainerCreationHelper::getContainerTableIndexDefinitions();
-        } else {
-            $indexes = $indexDefinitions;
-        }
-
-        $count = 1;
-        foreach($indexes as $tableName => $columns) {
-            try {
-                if(!$this->dbManager->createTableIndex($dbName, $count, $tableName, $columns)) {
-                    throw new GeneralException('Could not create database table indexes.');
-                }
-            } catch(AException $e) {
-                $this->logger->exception($e, __METHOD__);
-                continue;
-            }
-
-            $count++;
-        }
-    }
-
-    private function updateContainerDbSchema(string $dbName, string $containerId) {
-        try {
-            $conn = $this->dbManager->getConnectionToDatabase($dbName);
-        } catch(AException $e) {
-            throw new GeneralException('Could not establish connection to the container database.');
-        }
-
-        $sqlScripts = [];
-        $files = FileManager::getFilesInFolder(APP_ABSOLUTE_DIR . 'data\\db\\schema');
-        foreach($files as $file => $fileFullpath) {
-            if($file != "." && $file != "..") {
-                if(str_starts_with($file, 'u') && str_ends_with($file, '.php')) {
-                    $sqlScripts[$file] = $fileFullpath;
-                }
-            }
-        }
-
-        $this->logger->info('Found ' . count($sqlScripts) . ' DB schema updates.', __METHOD__);
-
-        foreach($sqlScripts as $scriptName => $script) {
-            $php = file_get_contents($script);
-
-            $schema = (int)(substr($scriptName, 1, -strlen('.php')));
-
-            $php = substr($php, strlen('<?php'));
-            $php = substr($php, 0, -strlen('?>'));
-
-            $result = eval($php);
-
-            if(!empty($result['tables'])) {
-                $this->createNewContainerTables($dbName, $result['tables']);
-            }
-            if(!empty($result['indexes'])) {
-                $this->createContainerTablesIndexes($dbName, $result['indexes']);
-            }
-            if(!empty($result['data'])) {
-                foreach($result['data'] as $tableName => $contents) {
-                    foreach($contents as $content) {
-                        $sql = 'INSERT INTO ' . $tableName . '(' . implode(', ', array_keys($content)) . ')';
-                        $sql .= ' VALUES (\'' . implode('\', \'', $content) . '\')';
-
-                        $conn->query($sql);
-                    }
-                }
-            }
-
-            $this->updateDbSchema($containerId, $schema);
-
-            $this->logger->info('Updated container database schema to ' . $schema . '.', __METHOD__);
-        }
-    }
-
-    private function updateDbSchema(string $containerId, int $schema) {
-        $sql = "UPDATE containers SET dbSchema = " . $schema . " WHERE containerId = '" . $containerId . "';";
-
-        $this->masterConn->query($sql);
-    }
-
-    public function checkContainerTitleExists(string $title) {
+    /**
+     * Checks if container's title exists
+     * 
+     * @param string $title Title
+     */
+    public function checkContainerTitleExists(string $title): bool {
         return $this->containerRepository->checkTitleExists($title);
     }
 
+    /**
+     * Changes container's status
+     * 
+     * @param string $containerId Container ID
+     * @param int $newStatus New status
+     * @param string $callingUserId Calling user ID
+     * @param string $description Description
+     */
     public function changeContainerStatus(string $containerId, int $newStatus, string $callingUserId, string $description) {
         $historyId = $this->createId(EntityManager::CONTAINER_STATUS_HISTORY);
 
@@ -490,6 +173,13 @@ class ContainerManager extends AManager {
         }
     }
 
+    /**
+     * Changes container's creation status
+     * 
+     * @param string $containerId Container ID
+     * @param int $percent Percent
+     * @param ?string $description Description
+     */
     public function changeContainerCreationStatus(string $containerId, int $percent, ?string $description) {
         $data = [
             'percentFinished' => (string)$percent,
@@ -501,6 +191,12 @@ class ContainerManager extends AManager {
         }
     }
 
+    /**
+     * Returns an instance of ContainerEntity for given container
+     * 
+     * @param string $containerId Container ID
+     * @param bool $force Force fetch data from the database (overrides cache and updates it)
+     */
     public function getContainerById(string $containerId, bool $force = false): ContainerEntity {
         $cache = $this->cacheFactory->getCache(CacheNames::CONTAINERS);
 
@@ -525,6 +221,12 @@ class ContainerManager extends AManager {
         return $containerEntity;
     }
 
+    /**
+     * Deletes container and all its data
+     * 
+     * @param string $containerId Container ID
+     * @param bool $isRequest False if it is a complete container or true if it is just a request
+     */
     public function deleteContainer(string $containerId, bool $isRequest = false) {
         $container = $this->getContainerById($containerId);
         
@@ -571,8 +273,16 @@ class ContainerManager extends AManager {
         foreach($databases as $database) {
             $this->containerDatabaseManager->dropDatabaseByEntryId($containerId, $database->getId());
         }
+
+        $this->cacheFactory->invalidateAllCache();
     }
 
+    /**
+     * Adds user to given container
+     * 
+     * @param string $userId User ID
+     * @param string $containerId Container ID
+     */
     public function addUserToContainer(string $userId, string $containerId) {
         $container = $this->getContainerById($containerId);
 
@@ -588,6 +298,12 @@ class ContainerManager extends AManager {
         $groupManager->addUserToGroupTitle(SystemGroups::ALL_USERS, $userId);
     }
 
+    /**
+     * Updates container
+     * 
+     * @param string $containerId Container ID
+     * @param array $data New data
+     */
     public function updateContainer(string $containerId, array $data) {
         if(!$this->containerRepository->updateContainer($containerId, $data)) {
             throw new GeneralException('Database error.');
@@ -598,7 +314,12 @@ class ContainerManager extends AManager {
         }
     }
 
-    public function getContainerUsageStatisticsTotalCount(string $containerId) {
+    /**
+     * Returns the number of all usage statistics entries for given container
+     * 
+     * @param string $containerId Container ID
+     */
+    public function getContainerUsageStatisticsTotalCount(string $containerId): ?int {
         $qb = $this->containerRepository->composeQueryForContainerUsageStatistics($containerId);
 
         $qb->select(['COUNT(entryId) AS count'])
@@ -607,6 +328,13 @@ class ContainerManager extends AManager {
         return $qb->fetch('count');
     }
 
+    /**
+     * Deletes container usage statistics
+     * 
+     * @param string $containerId Container ID
+     * @param int $limit Number of entries to delete
+     * @param bool $deleteAll True if all usage statistics for given container should be deleted
+     */
     public function deleteContainerUsageStatistics(string $containerId, int $limit, bool $deleteAll) {
         $entries = [];
 
@@ -627,10 +355,21 @@ class ContainerManager extends AManager {
         }
     }
 
+    /**
+     * Declines container request
+     * 
+     * @param string $containerId Container ID
+     */
     public function declineContainerRequest(string $containerId) {
         $this->deleteContainer($containerId, true);
     }
 
+    /**
+     * Approves container request and adds it to the container creation queue
+     * 
+     * @param string $container Container ID
+     * @param string $callingUserId Calling user ID
+     */
     public function approveContainerRequest(string $containerId, string $callingUserId) {
         $container = $this->getContainerById($containerId);
 
@@ -650,6 +389,29 @@ class ContainerManager extends AManager {
             !$this->cacheFactory->invalidateCacheByNamespace(CacheNames::NAVBAR_CONTAINER_SWITCH_USER_MEMBERSHIPS)) {
             throw new GeneralException('Could not invalidate cache.');
         }
+    }
+
+    /**
+     * Returns all containers that are in distribution
+     * 
+     * @return array<int, \App\Entities\ContainerEntity>
+     */
+    public function getContainersInDistribution(): array {
+        $qb = $this->containerRepository->composeQueryForContainers();
+        $qb->andWhere('isInDistribution = ?', ['1'])
+            ->execute();
+
+        $containerIds = [];
+        while($row = $qb->fetchAssoc()) {
+            $containerIds[] = $row['containerId'];
+        }
+
+        $containers = [];
+        foreach($containerIds as $containerId) {
+            $containers[] = $this->getContainerById($containerId, true);
+        }
+
+        return $containers;
     }
 }