+++ app/repositories/ARepository.php
@@ -30,7 +30,6 @@ abstract class ARepository extends AMultipleDatabaseConnectionHandler {
     public function __construct(DatabaseConnection $conn, Logger $logger) {
         parent::__construct($conn);
         $this->logger = $logger;
-        $this->cacheFactory = new CacheFactory();
 
         $this->transactionLogRepository = new TransactionLogRepository($this->conn, $this->logger);
     }