+++ app/ui/formbuilder2/FormBuilder2.php
@@ -3,6 +3,7 @@
 namespace App\UI\FormBuilder2;
 
 use App\Core\AjaxRequestBuilder;
+use App\Core\HashManager;
 use App\Core\Http\Ajax\Operations\CustomOperation;
 use App\Core\Http\Ajax\Operations\HTMLPageOperation;
 use App\Core\Http\Ajax\Requests\AAjaxRequest;
@@ -349,6 +350,24 @@ class FormBuilder2 extends AComponent {
         $fls = new FormLayoutSection($name);
 
         $this->elements[$name] = &$fls;
+    }
+
+    /**
+     * Adds easy security verification code check
+     * 
+     * @param string $name Element name
+     * @param ?string &$hash Already generated hash or implicitly generated hash
+     */
+    public function addSecurityVerificationCodeCheck(string $name, ?string &$hash) {
+        if($hash === null) {
+            $hash = HashManager::createHash(8);
+        }
+
+        $labelText = 'Enter this verification code below: <b>' . $hash . '</b>';
+
+        $this->addLabel('lbl_' . HashManager::createHash(4, false), $labelText);
+        $this->addTextInput($name, 'Verification code:')
+            ->setRequired();
     }
 
     /**