+++ CHANGELOG.md
@@ -20,6 +20,10 @@ Previously changelog for versions `1.0`-`1.4` has been located in README.md. How
     - Automatic running
     - Service scheduling
     - Service master & slave
+- Database migrations & initial seeding
+- Container distribution
+    - Containers can now be removed from distribution
+- Multiple bugfixes
 
 ## SkyDocu `1.3.1` (02/26/2025)
 - About application widget update